reorganized endpoint requests into files
